added -G:+CheckcastCounters to profile the paths taken through a compiled checkcast

+++ graal/com.oracle.max.cri/src/com/oracle/max/cri/xir/CiXirAssembler.java
@@ -871,7 +871,7 @@ public abstract class CiXirAssembler {
         return createConstant(CiConstant.forInt(v));
     }
 
-    public XirConstant l(int v) {
+    public XirConstant l(long v) {
         return createConstant(CiConstant.forLong(v));
     }
